cleanup
1. refs #1797

+++ hoot-rnd/src/main/cpp/hoot/rnd/io/ImplicitTagRulesSqliteReader.h
@@ -54,7 +54,7 @@ public:
 
   Tags getImplicitTags(const QSet<QString>& words, QSet<QString>& matchingWords);
 
-  int getRuleCount();
+  long getRuleCount();
 
 private:
